Implement following of association-lists
In order for a model-follower-cursor to follow a leader-cursor as it
traverses an association-list, association-fields have to be treated
as scalar fields, and the leader must not traverse the key-paths in
the associations. If the leader traverse key-paths, then the follower
will not be able to find the matching association to follow because
the entire key-path is needed to find the matching association.

So the model-leader-cursor has been updated to not traverse the key-
paths in associations. Now the model-follower-cursor is able to
follow the leader.

The model encoder had to be updated as a result of the above change.
It must now traverse the key-paths in associations.