Include type name in visitor method names
This helps avoid inference issues. When `LeaderManyModelVisitor` was
added, there were type inference issues with `Leader1ModelVisitor`
and so the former had to use type names in its method names. In this
commit, type names are added to the method names in other visitors
for the sake of consistency. But it also has the benefit of making it
easier to port this code to other languages that don't support method
overloading.