Make MainModel a subclass of SingleFieldModel
The main model composes the root entity like a single field composes an
entity. So this change makes them similar and allow common code for root
and non-root entities (the RootEntity class will be dropped in the next
commit). In the future, multiple main models with different meta-models
can be combined into a new main model since each main model is a field
and can be added to an entity as a field.