Parameterize entity matching to support null keys
A get-request model will contain null key values and will be used as a
follower-model in a forEach() traversal. These null keys need to match
any key value in the leader-model. To make this possible, and to support
other kinds of matching in the future, the follower-model-cursor can now
be supplied with a custom matching function that overrides the default
key matching in the model.