Stop using Buffer.peek() in the JSON tokenizer
The implementation peeks almost the entire JSON stream one character at
a time by calling Buffer.peek() each time (it has to be done each time
because the peek-buffer gets invalidated when the real buffer is read).
Buffer.peek() returns a new buffer instance every time it is called, so
getting rid of it is beneficial.