Drop generic types & constraints from operators
Generic types and constraints were attempted to ensure that operator
inputs are immutable and outputs are mutable, but to also ensure that
they are related generated model classes:

```
fun <I : MainModel, O> difference(
    oldModel: I,
    newModel: I,
    output: DifferenceModels<O>,
    mutableMainModelFactory: MutableMainModelFactory<O>
) where O : MutableMainModel, O : I {
```

Turns out that a compiler check has to be suppressed in order to achieve
the desired constraints, but does not work with Java-compatibility and
can be buggy at times [1].

There is not much benefit in using generic types and constraints if the
above constraints cannot be used. So they are being dropped, at least
until [2] is fixed.

[1] https://discuss.kotlinlang.org/t/bug-in-where/25011/2
[2] https://youtrack.jetbrains.com/issue/KT-19215