Revert getNewValue() to getOrNewValue()
getNewValue() was causing confusion for single fields and associations;
it gave the impression that the current value would be replaced by a new
value. So "OrNew" is being added back to the function name.

The only potential confusion with "OrNew" will be for Sets (and Lists,
but Lists will be dropped soon). But it is easier to reason that Sets
have many values and without specifying which value, the method won't
return an existing value, so it must be returning a new value. The
docs for the Set override of this method also explains this.